pyproject.toml
[project]
name = "agent-ivy"
version = "0.1.0"
description = "Telegram bot for news aggregation"
requires-python = ">=3.10"

dependencies = [
    "aiofiles==24.1.0",
    "aiogram==3.20.0.post0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.4",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "frozenlist==1.6.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "lxml==5.4.0",
    "magic-filter==1.0.12",
    "multidict==6.4.4",
    "packaging==25.0",
    "pendulum==3.1.0",
    "pluggy==1.6.0",
    "propcache==0.3.1",
    "pydantic==2.11.4",
    "pydantic_core==2.33.2",
    "pytest==8.3.5",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "requests==2.32.3",
    "six==1.17.0",
    "soupsieve==2.7",
    "typing_extensions==4.13.2",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "unshortenit==0.4.0",
    "urllib3==2.4.0",
    "yarl==1.20.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["bot*", "core*", "config*", "services*", "utils*"]
